requirements.txt
numpy<2
python-dotenv
opencv-python
peewee
ultralytics